+++ src/main/java/com/smartcoach/smartcoachBackend/Persistence/user/UsuarioClienteRepository.java
@@ -4,7 +4,7 @@ package com.smartcoach.smartcoachBackend.Persistence.user;
 import com.smartcoach.smartcoachBackend.Business.user.entities.UsuarioCliente;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UsuarioClienteRepository extends JpaRepository<UsuarioCliente, Integer> {
+public interface UsuarioClienteRepository extends JpaRepository<UsuarioCliente, Long> {
 }
